Add windows library to check file access
